Add eff field to ItemObject

+++ makepak/src/tomlinput.rs
@@ -80,7 +80,7 @@ pub struct ItemDepInput {
     pub basic_price: u32,
     pub gen_weight: f32,
     pub gen_level: u32,
-    pub eff: Option<i32>,
+    pub eff: Option<u16>,
     pub dice_n: Option<u16>,
     pub dice_x: Option<u16>,
     pub weapon_kind: Option<gamedata::item::WeaponKind>,
